requirements.txt
flask
tensorflow-cpu==2.11.0
Pillow
numpy
gunicorn